+++ server/include/ThreadPool.h
@@ -6,7 +6,7 @@
 
 #include <cstddef>
 #include <unordered_map>
-#include "Current.h"
+#include "Thread.h"
 #include "Server.h"
 
 class ThreadPool
